+++ docs/news.md
@@ -1,5 +1,14 @@
 # News
 
+## 0.1.2
+
+**Features:**
+
+- Version bump to latest version of `manifolds-rs`, unlocking GPU-accelerated
+  kNN searches.
+- GPU-accelerated version of EVoC that leverages the GPU-accelerated kNN
+  searches under the hood.
+
 ## 0.1.1
 
 **Features:**